Process notes into documentation
